FEAT (About) Change email to address
I noticed that the About page invited users to email me but the email address has "email Gareth J M Saunders" as the link text. It occurred to me that some users may prefer to write down the email address or at least see it clearly, e.g. on a mobile device.