FEAT (Homepage) Add abbr to SPB
SPB (Scottish Prayer Book) is not obvious to casual observers. Use the HTML abbr tag to add a hover tooltip to explain the abbreviation.

Use jQuery to search for instances of SPB in the text and dynamically replace this. This information is provided in a CSV file so it is impractical to mark that up as it is used for a variety of purposes.